Refactor userID type, add nested preloading
Refactored all of the user functions to use uint instead of uint64 for OS compatibility.

Added the helper function GetUserByID which obtains the user id from the JWT.

Refactored user database queries to use nested preloading. All queries are preloaded during the user data fetch which reduces page queries by 50%.

Added a dashboard route to the protecedAPI group.

Added a one-to-many association for inbox and sent messages to the user model.

+++ server/controller/category.go
@@ -37,7 +37,7 @@ func ViewCategory(context *gin.Context) {
 		return
 	}
 
-	category, err := model.GetCategoryByID(id)
+	category, err := model.GetCategoryByID(uint(id))
 
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
